src: flash: introduce microchip's envm flash driver
devices from Microchip's PolarFire SoC and PIC64GX families have on
board eNVM which is primarily used to store a bootloader. This patch
introduces a driver which utilises a loader algorithm to talk with the
on-board eNVM controller to write to the eNVM memory.

It does this by reading the sectors and offsets from a given elf file
and writing them to a working area in LIM (loosely integrated memory).
From there, the algorithm kicks in and writes the sections to the
eNVM.

Co-authored-by: Tommy Murphy <tommy_murphy@hotmail.com>
Co-authored-by: Anton Krug <anton.krug@gmail.com>
Co-authored-by: Daire McNamara <daire.mcnamara@microchip.com>
Co-authored-by: Matteo Bonicolini <matteo.bonicolini@microchip.com>
Signed-off-by: Conor Paxton <conor.paxton@microchip.com>